consolidate I/DownloadService
